+++ app/Http/Controllers/ProfileController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\ProfileUpdateRequest;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Support\Facades\Route;
 
 class ProfileController extends Controller
 {
@@ -14,10 +16,11 @@ class ProfileController extends Controller
      *
      * @return \Illuminate\View\View
      */
-    public function edit(Request $request)
+    public function edit(Request $request, User $profile)
     {
+        $this->authorize('update-profile', [$profile]);
         return view('profile.edit', [
-            'user' => $request->user(),
+            'user' => $profile
         ]);
     }
 
@@ -28,6 +31,7 @@ class ProfileController extends Controller
      */
     public function update(ProfileUpdateRequest $request)
     {
+        $this->authorize('update-profile', [User::find($request->route()->parameter('profile'))]);
         $request->user()->fill($request->validated());
 
         if ($request->user()->isDirty('email')) {
@@ -44,21 +48,40 @@ class ProfileController extends Controller
      *
      * @return \Illuminate\Http\RedirectResponse
      */
-    public function destroy(Request $request)
+    public function destroy(Request $request, User $profile)
     {
-        $request->validateWithBag('userDeletion', [
-            'password' => ['required', 'current-password'],
-        ]);
-
-        $user = $request->user();
-
-        Auth::logout();
+        $this->authorize('delete-profile', [$profile]);
+        if (!$request->user()->is_admin) {
+            $request->validateWithBag('userDeletion', [
+                'password' => ['required', 'current-password'],
+            ]);
+        }
+        $user = $profile;
+        if ($request->user()->id == $profile->id) {
+            Auth::logout();
+        }
 
         $user->delete();
 
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-
+        if ($request->user()->id == $profile->id) {
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
+        }
+        $profile = null;
         return Redirect::to('/');
+    }
+
+    public function block(User $profile)
+    {
+        $profile->blocked = 1;
+        $profile->save();
+        return back();
+    }
+
+    public function unblock(User $profile)
+    {
+        $profile->blocked = 0;
+        $profile->save();
+        return back();
     }
 }
